check if .env exists

+++ test/test_helper.rb
@@ -9,18 +9,20 @@ class ActiveSupport::TestCase
   # -- they do not yet inherit this setting
   fixtures :all
 
-  File.open(".env", "r").each_line do |line|
-  	key = ""
-  	value = ""
-  	on_key = true
-  	line.split("=").map.each { |token|
-  	  if on_key
-  	  	key = token
-  	  else
-  	  	value = token.gsub(/"/, "").chomp
-  	  end
-  	  on_key = false
-  	}
-  	ENV[key] = value if key != "RAILS_ENV"
+  if File.exist?(".env")
+    File.open(".env", "r").each_line do |line|
+    	key = ""
+    	value = ""
+    	on_key = true
+    	line.split("=").map.each { |token|
+    	  if on_key
+    	  	key = token
+    	  else
+    	  	value = token.gsub(/"/, "").chomp
+    	  end
+    	  on_key = false
+    	}
+    	ENV[key] = value if key != "RAILS_ENV"
+    end
   end
 end
